fix: bug with new cell appearance
There was bug with new cell appearance after turn - when player click
hotkey but there is no empty space in this destination new cell appeared
anyway. That's was wrong behavior
Added isCellShifted field to shift module, update it at each successful
shift and return in the end of all iterations
Use isCellShifted value in Game component to determine should game render
new cell or not

Find additional bug when code in shift module in if Else block run on
each iteration. Added additional check of passed value to fix it